+++ src/lib/components/modal/contract/SaveNewContract.tsx
@@ -9,6 +9,7 @@ import { ActionModal } from "../ActionModal";
 import {
   useBaseApiRoute,
   useCelatoneApp,
+  useExampleAddresses,
   useValidateAddress,
 } from "lib/app-provider";
 import type { FormStatus } from "lib/components/forms";
@@ -47,11 +48,7 @@ export function SaveNewContractModal({
   const { getContractLocalInfo } = useContractStore();
   const { validateContractAddress } = useValidateAddress();
 
-  const {
-    chainConfig: {
-      exampleAddresses: { contract: exampleContractAddress },
-    },
-  } = useCelatoneApp();
+  const { contract: exampleContractAddress } = useExampleAddresses();
   const initialList =
     list.value === formatSlugName(INSTANTIATED_LIST_NAME) ? [] : [list];