+++ src/lib/pages/instantiate/instantiate.tsx
@@ -10,11 +10,11 @@ import { useForm } from "react-hook-form";
 import {
   useFabricateFee,
   useInstantiateTx,
-  useCelatoneApp,
   useValidateAddress,
   useSimulateFeeQuery,
   useCurrentChain,
   useBaseApiRoute,
+  useExampleAddresses,
 } from "lib/app-provider";
 import { AssignMe } from "lib/components/AssignMe";
 import { ConnectWalletAlert } from "lib/components/ConnectWalletAlert";
@@ -69,11 +69,8 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
   const router = useRouter();
   const msgQuery = (router.query.msg as string) ?? "";
   const codeIdQuery = (router.query["code-id"] as string) ?? "";
-  const {
-    chainConfig: {
-      exampleAddresses: { contract: exampleContractAddress },
-    },
-  } = useCelatoneApp();
+  const { user: exampleUserAddress } = useExampleAddresses();
+
   const { address = "" } = useCurrentChain();
   const lcdEndpoint = useBaseApiRoute("rest");
 
@@ -338,7 +335,7 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
             name="adminAddress"
             control={control}
             label="Admin Address (optional)"
-            placeholder={`ex. ${exampleContractAddress}`}
+            placeholder={`ex. ${exampleUserAddress}`}
             helperText="The contract's admin will be able to migrate and update future admins."
             variant="floating"
             error={validateAdmin(adminAddress)}